Fixed crash from exp updates + Fixed some guild stuff + Changed code in fillSpawnStruct to fix spawn HP percentages + Fixed player HP data + Updated hpNpcUpdateStruct + Updated version to 5.13.10.10
git-svn-id: https://svn.code.sf.net/p/seq/svn/showeq/trunk@782 19a12193-7728-0410-8f54-e9f61ebc0c9f

+++ src/player.cpp
@@ -794,7 +794,7 @@ void Player::updateStamina(const uint8_t* data)
     savePlayerState();
 }
 
-void Player::setLastKill(const QString& name, uint8_t level)
+void Player::setLastKill(const QString& name, int level)
 {
   // note the last spawn this player killed
   m_lastSpawnKilledName = name;
@@ -1234,7 +1234,7 @@ uint16_t Player::deity() const
 	  m_defaultDeity : m_deity); 
 }
 
-uint8_t Player::level() const 
+int Player::level() const
 { 
   return (!m_useAutoDetectedSettings || m_useDefaults ? 
 	  m_defaultLevel : m_level);
